Move conv code into different files
