+++ lib/nnc/cpu/conv/ccv_nnc_cmd_conv_opt.h
@@ -1,0 +1,19 @@
+/**********************************************************
+ * C-based/Cached/Core Computer Vision Library
+ * Liu Liu, 2010-02-01
+ **********************************************************/
+
+/**********************************************************
+ * CCV - Neural Network Collection
+ **********************************************************/
+
+#ifndef GUARD_ccv_nnc_cmd_conv_opt_h
+#define GUARD_ccv_nnc_cmd_conv_opt_h
+
+#include <ccv.h>
+#include <nnc/ccv_nnc.h>
+
+int ccv_nnc_conv_forw_4x4_3x3_winograd(const ccv_nnc_tensor_view_t* a, const ccv_nnc_tensor_t* w, const ccv_nnc_tensor_t* bias, const ccv_nnc_hint_t hint, ccv_nnc_tensor_view_t* b);
+int ccv_nnc_conv_forw_opt(const ccv_nnc_tensor_view_t* a, const ccv_nnc_tensor_t* w, const ccv_nnc_tensor_t* bias, const ccv_nnc_hint_t hint, ccv_nnc_tensor_view_t* b);
+
+#endif
